prgtest 52_30

+++ 52_30_reanomena/Lloguer.java
@@ -6,15 +6,15 @@
 public class Lloguer {
     private int dies;
     private Vehicle vehicle;
-    private static final double QUANTITAT_BASIC = 3;
-    private static final double DIES_BASIC = 3;
+    private static final int QUANTITAT_BASIC = 3;
+    private static final int DIES_BASIC = 3;
     private static final double MULT_BASIC = 1.5;
 
-    private static final double QUANTITAT_GENERAL = 4;
-    private static final double DIES_GENERAL = 2;
+    private static final int QUANTITAT_GENERAL = 4;
+    private static final int DIES_GENERAL = 2;
     private static final double MULT_GENERAL = 2.5;
 
-    private static final double QUANTITAT_LUXE = 6;
+    private static final int QUANTITAT_LUXE = 6;
 
     Lloguer(Vehicle vehicle, int dies) {
         this.dies = dies;
